fisher, force, gator

+++ src/e-k/Gator/Gator.Models.cs
@@ -5,8 +5,8 @@ namespace Skender.Stock.Indicators
     [Serializable]
     public class GatorResult : ResultBase
     {
-        public decimal? Upper { get; set; }
-        public decimal? Lower { get; set; }
+        public double? Upper { get; set; }
+        public double? Lower { get; set; }
 
         public bool? UpperIsExpanding { get; set; }
         public bool? LowerIsExpanding { get; set; }
